@@ -7,9 +7,7 @@ func (si *searchInfo) evaluate(pos *chess.Position) int32 {
 	player := pos.Turn()
 	knights, bishops, rooks, queens := pos.CountPieces()
 	phase := int32(knights + bishops + 2*rooks + 4*queens)
-	if phase > 24 {
-		phase = 24 // in case of early promotion
-	}
+	phase = min(phase, 24) // in case of early promotion
 
 	var mgMaterial [2]int32
 	var egMaterial [2]int32
